fix(modules): `storage` is additive, so it costs no schemaVersion
A version bump is for a change that BREAKS a reader, and this is not one: a
reader that predates `storage` sees a module with no database, which is exactly
what such a module had. The renames that made v2 necessary were a different
thing entirely -- `dependsOn` parsed as absent and a module installed with its
dependencies silently dropped.

What a new bundle actually needs is a newer HOST, because its binary resolves
sessions through `/_host/session` and expects its grant in the spawn env. That
is `engines`, which every manifest now floors at the server carrying both. So
v3 goes away, `storage` joins v2 where it belongs, and the contract stays at 2.
