+++ modules/tv.kroma.engine.transmission/module.json
@@ -1,12 +1,12 @@
 {
-  "$schema": "https://modules.kroma.tv/schemas/3/manifest.json",
-  "schemaVersion": 3,
+  "$schema": "https://modules.kroma.tv/schemas/2/manifest.json",
+  "schemaVersion": 2,
   "id": "tv.kroma.engine.transmission",
   "name": "Transmission engine",
   "version": "0.2.0",
   "description": "Transmission RPC as a download sub-engine for the Downloads module. Toggle it to add/remove Transmission from the download-client picker.",
   "engines": {
-    "server": ">=0.1.4"
+    "server": ">=0.1.38"
   },
   "dependencies": {
     "tv.kroma.torrents": "^0.2.0"
