+++ modules/tv.kroma.scene/module.json
@@ -1,12 +1,12 @@
 {
-  "$schema": "https://modules.kroma.tv/schemas/3/manifest.json",
-  "schemaVersion": 3,
+  "$schema": "https://modules.kroma.tv/schemas/2/manifest.json",
+  "schemaVersion": 2,
   "id": "tv.kroma.scene",
   "name": "Release parser",
   "version": "0.2.0",
   "description": "Scene / P2P release-name parsing and scoring for the acquisition stack.",
   "engines": {
-    "server": ">=0.1.4"
+    "server": ">=0.1.38"
   },
   "library": true,
   "provides": [
