+++ modules/tv.kroma.torznab/module.json
@@ -1,12 +1,12 @@
 {
-  "$schema": "https://modules.kroma.tv/schemas/3/manifest.json",
-  "schemaVersion": 3,
+  "$schema": "https://modules.kroma.tv/schemas/2/manifest.json",
+  "schemaVersion": 2,
   "id": "tv.kroma.torznab",
   "name": "Torznab indexers",
   "version": "0.2.0",
   "description": "External Torznab / Newznab indexers (Jackett, Prowlarr).",
   "engines": {
-    "server": ">=0.1.4"
+    "server": ">=0.1.38"
   },
   "provides": [
     {
