+++ packages/registry/src/manifest/v2.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { dependenciesOf, Manifest, optionalDependenciesOf } from './v2.ts';
+import { dependenciesOf, Manifest, optionalDependenciesOf, speaksCurrentSchema } from './v2.ts';
 
 const base = { id: 'tv.kroma.x', name: 'X', version: '1.0.0' };
 
@@ -32,6 +32,52 @@ describe('Manifest', () => {
   });
 });
 
+describe('storage', () => {
+  it('is absent for a module that declares none, which is most of them', () => {
+    // Absent is not the same as empty: it means no database at all, and a
+    // sidecar built for it does not link SQLite.
+    expect(Manifest.parse(base).storage).toBeUndefined();
+  });
+
+  it('reads an empty object as the module’s own file and nothing more', () => {
+    const parsed = Manifest.parse({ ...base, storage: {} });
+    expect(parsed.storage).toEqual({});
+    expect(parsed.storage?.core).toBeUndefined();
+  });
+
+  it('keeps the core grant whole, table and column entries alike', () => {
+    const parsed = Manifest.parse({
+      ...base,
+      storage: {
+        core: { read: ['requests', 'users.username'], write: ['wanted'] },
+        adopt: ['indexers'],
+      },
+    });
+    // Dropping any of this would widen or silently void a module's grant.
+    expect(parsed.storage).toEqual({
+      core: { read: ['requests', 'users.username'], write: ['wanted'] },
+      adopt: ['indexers'],
+    });
+  });
+
+  it('refuses an entry that is not a table or one of its columns', () => {
+    const bad = (read: string[]) => () => Manifest.parse({ ...base, storage: { core: { read } } });
+    expect(bad(['users.name.first'])).toThrow();
+    expect(bad(['requests; DROP TABLE users'])).toThrow();
+    expect(bad(['*'])).toThrow();
+    expect(bad([''])).toThrow();
+    // ...and an adopted table is a bare name, never a column.
+    expect(() => Manifest.parse({ ...base, storage: { adopt: ['users.id'] } })).toThrow();
+  });
+
+  it('is additive: a manifest that predates it still parses', () => {
+    // Which is why it did not cost a schemaVersion. A reader that does not know
+    // the field sees a module with no database, and that is what it had.
+    expect(speaksCurrentSchema({ schemaVersion: 2 })).toBe(true);
+    expect(Manifest.parse({ ...base, schemaVersion: 2 }).storage).toBeUndefined();
+  });
+});
+
 describe('the shape a real manifest declares', () => {
   // Every property a manifest may carry. zod STRIPS what it does not declare,
   // so a field missing here is a field the published catalog would lose without
@@ -52,6 +98,10 @@ describe('the shape a real manifest declares', () => {
     permissions: ['library.manage'],
     config: [{ key: 'host', label: 'Host', type: 'string', placeholder: 'localhost' }],
     feRemote: { module: './Module' },
+    storage: {
+      core: { read: ['downloads', 'requests'], write: ['downloads'] },
+      adopt: ['download_clients'],
+    },
   };
 
   it('keeps every one of them', () => {
