+++ packages/registry/src/manifest/v2.ts
@@ -6,7 +6,6 @@
 
 import { z } from 'zod';
 import { Capability, CapabilityReq } from '../documents/v1.ts';
-import { speaksSchema } from './shared.ts';
 
 // Any optional field may arrive as an explicit null.
 const DependencyMap = z.record(z.string(), z.string()).nullish();
@@ -30,6 +29,41 @@ const SCHEMA_VERSION = 2;
 
 /** A module id: reverse-DNS, lowercase. */
 export const REVERSE_DNS_ID = /^[a-z0-9]+(?:\.[a-z0-9-]+)+$/;
+
+// A table (`downloads`) or one of its columns (`users.username`). Anything not
+// listed is denied, so the spelling has to be exact.
+const TableOrColumn = z.string().regex(/^[A-Za-z_]\w*(?:\.[A-Za-z_]\w*)?$/);
+
+/** The slice of the SHARED core database a module may reach. */
+export const CoreScope = z.object({
+  read: z
+    .array(TableOrColumn)
+    .nullish()
+    .describe(
+      'Tables and columns this module may SELECT, as "table" or "table.column". A column named in a WHERE is reached as much as one that is projected.',
+    ),
+  write: z
+    .array(TableOrColumn)
+    .nullish()
+    .describe(
+      'Tables this module may INSERT / UPDATE / DELETE. A write grant is not a read grant, and a foreign key drags its other table in: writing a child reads the parent, and a cascading delete writes the child.',
+    ),
+});
+export type CoreScope = z.infer<typeof CoreScope>;
+
+/** A module's databases, and the capability itself. */
+export const Storage = z.object({
+  core: CoreScope.nullish().describe(
+    "The slice of the shared core database this module may reach. Enforced per connection by SQLite's authorizer at prepare time, so it cannot be worked around by building the SQL as a string. Absent means none of it.",
+  ),
+  adopt: z
+    .array(z.string().regex(/^[A-Za-z_]\w*$/))
+    .nullish()
+    .describe(
+      "Tables this module used to keep in the core database and now owns. The core moves each one - schema, indexes and rows - into the module's own file before the module is spawned, then drops the core copy, so it happens exactly once.",
+    ),
+});
+export type Storage = z.infer<typeof Storage>;
 
 /** A module's `module.json`, as authored. */
 export const Manifest = z.object({
@@ -109,6 +143,9 @@ export const Manifest = z.object({
     .array(ConfigField)
     .nullish()
     .describe('Admin-configurable settings this module exposes, rendered as a form.'),
+  storage: Storage.nullish().describe(
+    "This module's databases, and the capability itself: without a storage object the module gets no database and its binary does not link SQLite. Its presence alone grants it its own file at <data>/modules/<id>/module.sqlite, where its migrations run and which it owns outright.",
+  ),
   feRemote: z
     .object({ module: z.string() })
     .nullish()
@@ -132,10 +169,25 @@ export const Manifest = z.object({
 });
 export type Manifest = z.infer<typeof Manifest>;
 
-/** Whether a module was built against the manifest contract this file defines. */
-export const speaksCurrentSchema = speaksSchema(SCHEMA_VERSION);
+/** Whether a module was built against the manifest contract this build speaks.
+ *  A bundle that was not is refused rather than read on a best-effort basis: the
+ *  fields that moved between versions parse as ABSENT, not as errors, so a stale
+ *  one would install with its dependencies silently dropped. */
+export function speaksCurrentSchema(manifest: Pick<Manifest, 'schemaVersion'>): boolean {
+  return manifest.schemaVersion === SCHEMA_VERSION;
+}
 
-export { dependenciesOf, optionalDependenciesOf } from './shared.ts';
+const mapOf = (raw: Manifest['dependencies']): Record<string, string> => raw ?? {};
+
+/** The versions a module requires, by id. */
+export function dependenciesOf(manifest: Manifest): Record<string, string> {
+  return mapOf(manifest.dependencies);
+}
+
+/** The versions a module suggests but does not require, by id. */
+export function optionalDependenciesOf(manifest: Manifest): Record<string, string> {
+  return mapOf(manifest.optionalDependencies);
+}
 
 /** What a registry document says about one downloadable build. */
 export const ArtifactRef = z.object({
