+++ server/src/api/it_admin2.rs
@@ -241,15 +241,15 @@ async fn store_catalog_enriches_a_reachable_registry() {
     let t = test_app();
     let url = spawn_registry(json!([
             // Universal (library) bundle: satisfies any host -> compatible.
-            { "schemaVersion": 3, "id": "tv.kroma.lib", "name": "Lib", "version": "1.0.0", "library": true,
+            { "schemaVersion": 2, "id": "tv.kroma.lib", "name": "Lib", "version": "1.0.0", "library": true,
               "dependencies": { "tv.kroma.dep": "^0.1.0" },
               "artifacts": [{ "target": null, "url": "https://x/lib.kmod", "size": 10, "sha256": "aa" }] },
             // An engine range no server satisfies -> incompatible with a reason.
-            { "schemaVersion": 3, "id": "tv.kroma.future", "name": "Future", "version": "2.0.0",
+            { "schemaVersion": 2, "id": "tv.kroma.future", "name": "Future", "version": "2.0.0",
               "engines": { "server": ">=999.0.0" },
               "artifacts": [{ "target": null, "url": "https://x/future.kmod" }] },
             // Only a foreign-arch build -> no artifact for this platform.
-            { "schemaVersion": 3, "id": "tv.kroma.alien", "name": "Alien", "version": "1.0.0",
+            { "schemaVersion": 2, "id": "tv.kroma.alien", "name": "Alien", "version": "1.0.0",
               "artifacts": [{ "target": "sparc-unknown-none-elf", "url": "https://x/alien.kmod" }] }
         ]))
     .await;
@@ -322,7 +322,7 @@ async fn the_store_reads_an_rfc_110_registry_through_its_descriptor() {
             "modules": ["com.acme.app"]
         }),
         json!([
-            { "schemaVersion": 3, "id": "com.acme.app", "name": "App", "version": "1.2.0", "library": true,
+            { "schemaVersion": 2, "id": "com.acme.app", "name": "App", "version": "1.2.0", "library": true,
               "dependencies": { "com.acme.lib": "^1.0.0" },
               "optionalDependencies": { "com.acme.vpn": "*" },
               "artifacts": [{ "target": null, "url": "https://x/app.kmod", "size": 10,
@@ -350,7 +350,7 @@ async fn a_registry_speaking_a_contract_this_server_does_not_know_is_not_read() 
     let t = test_app();
     let url = spawn_rfc_registry(
         json!({ "apiVersion": 99, "name": "Future", "url": "https://f", "modules": ["com.acme.app"] }),
-        json!([{ "schemaVersion": 3, "id": "com.acme.app", "name": "App", "version": "1.2.0", "artifacts": [] }]),
+        json!([{ "schemaVersion": 2, "id": "com.acme.app", "name": "App", "version": "1.2.0", "artifacts": [] }]),
     )
     .await;
     point_store_at(&t, &url);
@@ -372,16 +372,16 @@ fn point_store_at(t: &crate::api::test_support::TestApp, url: &str) {
 async fn the_install_plan_pulls_dependencies_first_and_offers_the_rest() {
     let t = test_app();
     let url = spawn_registry(json!([
-            { "schemaVersion": 3, "id": "tv.x.app", "name": "App", "version": "1.0.0",
+            { "schemaVersion": 2, "id": "tv.x.app", "name": "App", "version": "1.0.0",
               "dependencies": { "tv.x.lib": "^1.0.0" },
               "optionalDependencies": { "tv.x.vpn": "*" },
               "requires": [{ "kind": "download-client" }],
               "artifacts": [{ "target": null, "url": "https://x/app.kmod", "size": 30 }] },
-            { "schemaVersion": 3, "id": "tv.x.lib", "name": "Lib", "version": "1.4.0", "library": true,
+            { "schemaVersion": 2, "id": "tv.x.lib", "name": "Lib", "version": "1.4.0", "library": true,
               "artifacts": [{ "target": null, "url": "https://x/lib.kmod", "size": 12 }] },
-            { "schemaVersion": 3, "id": "tv.x.vpn", "name": "VPN", "version": "1.0.0",
+            { "schemaVersion": 2, "id": "tv.x.vpn", "name": "VPN", "version": "1.0.0",
               "artifacts": [{ "target": null, "url": "https://x/vpn.kmod", "size": 7 }] },
-            { "schemaVersion": 3, "id": "tv.x.engine", "name": "Engine", "version": "1.0.0",
+            { "schemaVersion": 2, "id": "tv.x.engine", "name": "Engine", "version": "1.0.0",
               "provides": [{ "kind": "download-client", "id": "rqbit" }],
               "artifacts": [{ "target": null, "url": "https://x/engine.kmod", "size": 5 }] }
         ]))
@@ -393,7 +393,7 @@ async fn the_install_plan_pulls_dependencies_first_and_offers_the_rest() {
         "POST",
         "/api/admin/store/plan",
         Some(&t.token),
-        Some(json!({ "schemaVersion": 3, "id": "tv.x.app" })),
+        Some(json!({ "schemaVersion": 2, "id": "tv.x.app" })),
     )
     .await;
     assert_eq!(status, StatusCode::OK);
@@ -421,7 +421,7 @@ async fn the_install_plan_pulls_dependencies_first_and_offers_the_rest() {
         "POST",
         "/api/admin/store/plan",
         Some(&t.token),
-        Some(json!({ "schemaVersion": 3, "id": "tv.x.app", "include": ["tv.x.app", "tv.x.vpn"] })),
+        Some(json!({ "schemaVersion": 2, "id": "tv.x.app", "include": ["tv.x.app", "tv.x.vpn"] })),
     )
     .await;
     assert_eq!(status, StatusCode::OK);
@@ -442,7 +442,7 @@ async fn planning_a_module_the_registry_does_not_offer_is_a_client_error() {
         "POST",
         "/api/admin/store/plan",
         Some(&t.token),
-        Some(json!({ "schemaVersion": 3, "id": "tv.x.ghost" })),
+        Some(json!({ "schemaVersion": 2, "id": "tv.x.ghost" })),
     )
     .await;
     assert_eq!(status, StatusCode::BAD_REQUEST);
@@ -459,7 +459,7 @@ async fn an_official_registry_that_is_down_stops_a_plan_rather_than_offering_les
         "POST",
         "/api/admin/store/plan",
         Some(&t.token),
-        Some(json!({ "schemaVersion": 3, "id": "tv.kroma.torrents" })),
+        Some(json!({ "schemaVersion": 2, "id": "tv.kroma.torrents" })),
     )
     .await;
     assert_eq!(status, StatusCode::BAD_REQUEST);
@@ -470,7 +470,7 @@ async fn an_official_registry_that_is_down_stops_a_plan_rather_than_offering_les
 async fn an_added_registry_that_is_down_leaves_the_official_catalog_usable() {
     let t = test_app();
     let url = spawn_registry(json!([
-            { "schemaVersion": 3, "id": "tv.x.app", "name": "App", "version": "1.0.0",
+            { "schemaVersion": 2, "id": "tv.x.app", "name": "App", "version": "1.0.0",
               "artifacts": [{ "target": null, "url": "https://x/app.kmod", "size": 3 }] }
         ]))
     .await;
@@ -493,7 +493,7 @@ async fn an_added_registry_that_is_down_leaves_the_official_catalog_usable() {
         "POST",
         "/api/admin/store/plan",
         Some(&t.token),
-        Some(json!({ "schemaVersion": 3, "id": "tv.x.app" })),
+        Some(json!({ "schemaVersion": 2, "id": "tv.x.app" })),
     )
     .await;
     assert_eq!(status, StatusCode::OK);
@@ -517,7 +517,7 @@ async fn the_install_plan_is_closed_to_anyone_who_cannot_manage_the_server() {
         "POST",
         "/api/admin/store/plan",
         Some(&m),
-        Some(json!({ "schemaVersion": 3, "id": "tv.x.app" })),
+        Some(json!({ "schemaVersion": 2, "id": "tv.x.app" })),
     )
     .await;
     assert_eq!(status, StatusCode::FORBIDDEN);
@@ -648,7 +648,7 @@ async fn an_official_catalog_that_reports_its_own_outage_is_a_failure_not_an_emp
         "POST",
         "/api/admin/store/plan",
         Some(&t.token),
-        Some(json!({ "schemaVersion": 3, "id": "tv.kroma.torrents" })),
+        Some(json!({ "schemaVersion": 2, "id": "tv.kroma.torrents" })),
     )
     .await;
     assert_eq!(status, StatusCode::BAD_REQUEST);
